Fixed save and load
Now it needs to be implemented in a way that it can save to files and 
load from them. The current implementation is dependant on that data-id 
property. I'd like to remove that if possible. It's possible to just 
save "grid.getItems()" to a variable, and then to sort by said variable, 
but the variable can't be stringified.