Save to and load from files
Also implemented autosaving that works for the desktop app and 
in-browser. The saving works by retrieving the inner HTML elements of 
all the cards in the grid and then saves an array as a string. Loading 
the file is as simple as parsing the string and then adding all the 
items back using an item template. There's currently a bug where 
uninstalling Sieve doesn't clear the autosaved data. I have yet to test 
how this works on macOS and Linux.