Tweaked card adding
Tweaked card adding so that the template's border colour updates 
whenever the ghost is clicked. This fixes an issue where clicking regBtn 
would set the template's colour to be that of the last card, and then if 
you went out and dragged a different grid to be the last one, the 
template would still be the previous group's colour instead of the new 
last one's.